+++ frontend/src/pages/RefreshTokenHandler/RefreshTokenHandler.tsx
@@ -3,12 +3,12 @@ import { useEffect } from 'react';
 import { useQuery } from 'react-query';
 
 import { getAccessTokenByRefreshToken } from '@/api/login';
+import LoginLoading from '@/components/login/LoginLoading/LoginLoading';
 import { ErrorMessage } from '@/constants/ErrorMessage';
 import { ACCESSTOKEN_KEY } from '@/constants/index';
 import { URL } from '@/constants/url';
 import useSnackBar from '@/hooks/common/useSnackBar';
 import useThrowCustomError from '@/hooks/common/useThrowCustomError';
-import LoginLoading from '@/pages/Login/LoginLoading/LoginLoading';
 
 const RefreshTokenHandler = () => {
 	const { data, isSuccess, isError, error } = useQuery<
